Update entities columns to hold more than 64 entries
Senses entities (pos, misc, dial and field) were limited to 64 per type
due to the use of a 64-bit mask to code them. However pos has grown over
64 since quite some time already.

This patch adds columns as needed when we create the database (pos
becomes pos0 and pos1), and renames the columns through the program from
foo to foo0. No other change is made so entity querying is still buggy.